Use PortsScanned method for summary output
Use earlier output from rs.PortsScanned() vs trying to
calculate on the spot the count of entries in the results
set to determine the number of ports scanned.

Right now the two produce the same results, but using the
method is likely to continue giving the expected results
after any future refactoring.