HOTFIX | Apply exact path match for peers list
This is not the long-term fix, but for now this commit
updates the xpath expression to match *exactly* the
XML document structure used in non-group-based LOCKSS
networks to define peer nodes.

This change causes the application to "fail fast" when
used on a LOCKSS node which *does* participate in a
network which uses group-based peers lists.

refs GH-21